role based access control

+++ src/services/productService.js
@@ -9,6 +9,13 @@ const getProducts = async (query) => {
 const getProductById = async (id) => {
   const product = await Product.findById(id);
 
+  if (!product) {
+    throw {
+      statusCode: 404,
+      message: "Product not found.",
+    };
+  }
+
   return product;
 };
 
@@ -21,7 +28,16 @@ const createProduct = async (data, createdBy) => {
   return createdProduct;
 };
 
-const updateProduct = async (id, data) => {
+const updateProduct = async (id, data, userId) => {
+  const product = await getProductById(id);
+
+  if (product.createdBy != userId) {
+    throw {
+      statusCode: 403,
+      message: "Access denied.",
+    };
+  }
+
   const updatedProduct = await Product.findByIdAndUpdate(id, data, {
     new: true,
   });
@@ -29,7 +45,16 @@ const updateProduct = async (id, data) => {
   return updatedProduct;
 };
 
-const deleteProduct = async (id) => {
+const deleteProduct = async (id, userId) => {
+  const product = await getProductById(id);
+
+  if (product.createdBy != userId) {
+    throw {
+      statusCode: 403,
+      message: "Access denied.",
+    };
+  }
+
   await Product.findByIdAndDelete(id);
 };
